+++ dlib/CMakeLists.txt
@@ -839,9 +839,6 @@ if (NOT TARGET dlib)
          if (FFMPEG_FOUND)
             list (APPEND dlib_needed_public_includes ${FFMPEG_INCLUDE_DIRS})
             list (APPEND dlib_needed_public_libraries ${FFMPEG_LINK_LIBRARIES})
-            set(source_files ${source_files}
-               media/ffmpeg_utils.cpp
-               media/ffmpeg_demuxer.cpp)
             enable_preprocessor_switch(DLIB_USE_FFMPEG)
          else()
             set(DLIB_USE_FFMPEG OFF CACHE BOOL ${DLIB_USE_FFMPEG_STR} FORCE )
